auth middleware on every protected route

+++ server/src/albums/albums.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { HttpException, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { Model } from 'mongoose'
 import { Album } from 'src/models/Album.schema'
@@ -8,23 +8,32 @@ import { CreateOrUpdateAlbumDto } from './dto/albums.dto'
 export class AlbumsService {
   constructor(@InjectModel(Album.name) private albumModel: Model<Album>) {}
 
-  getMyAlbums() {
-    return this.albumModel.find({})
+  getMyAlbums(userId: string) {
+    return this.albumModel.find({ owner: userId })
   }
 
-  getAlbumById(id: string) {
-    return this.albumModel.findById(id)
+  async getAlbumById(id: string, userId: string) {
+    const album = await this.albumModel.findOne({ _id: id, owner: userId })
+
+    if (!album) {
+      throw new HttpException('Album Not Found!', 404)
+    }
+    return album
   }
 
-  async createNewAlbum(data: CreateOrUpdateAlbumDto, owner: string) {
-    const album = new this.albumModel({ ...data, owner })
+  async createNewAlbum(data: CreateOrUpdateAlbumDto, userId: string) {
+    const album = new this.albumModel({ ...data, owner: userId })
     await album.save()
 
     return album
   }
 
-  async updateAlbum(id: string, data: CreateOrUpdateAlbumDto) {
-    const album = await this.albumModel.findById(id)
+  async updateAlbum(id: string, data: CreateOrUpdateAlbumDto, userId: string) {
+    const album = await this.albumModel.findOne({ _id: id, owner: userId })
+
+    if (!album) {
+      throw new HttpException('Album Not Found!', 404)
+    }
 
     Object.keys(data).forEach((e) => {
       album[e] = data[e]
@@ -35,8 +44,13 @@ export class AlbumsService {
     return album
   }
 
-  async deleteAlbum(id: string) {
-    const album = await this.albumModel.findById(id)
+  async deleteAlbum(id: string, userId: string) {
+    const album = await this.albumModel.findOne({ _id: id, owner: userId })
+
+    if (!album) {
+      throw new HttpException('Album Not Found!', 404)
+    }
+
     await album.delete()
 
     return album
